+++ packages/executor/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @midl-xyz/midl-js-executor
 
+## 2.0.0-next.26
+
+### Patch Changes
+
+- ccc957e: feat: use btc fee rate from midl
+
+## 2.0.0-next.25
+
+### Patch Changes
+
+- fix(executor): fee rate multiplier
+
 ## 2.0.0-next.24
 
 ### Patch Changes